platform: introduce wrapper for the memfd_create syscall
For an inflight fuctionality a new VHOST_USER_GET_INFLIGHT_FD command
should be supported. vhost-server daemon should allocate memory and send
a file descriptor back to the master. The easiest way to do it in Linux
is by using memfd_create syscall.
Since we are using the ubuntu xenial distribution with glibc-2.23 we
should add our wrapper for this syscall. Add the memfd_syscall wrapper
to platform for future usage.
Remove the _POSIX_C_SOURCE=200809L definition since it conflicts with
the syscall declaration.

Signed-off-by: Dima Stepanov <dimastep@yandex-team.ru>