vdev: base support for GET_INFLIGHT_FD and SET_INFLIGHT_FD
Add new vhost-user commands:
  VHOST_USER_GET_INFLIGHT_FD
  VHOST_USER_SET_INFLIGHT_FD
This is just a basic support to handle these commands. The GET commands
allocates a buffer according to vhost-user documentation and send back
FD. The SET command receives an FD and update its structures. For now
only the split vrings are supported. The following structures are added:
  vhost_user_inflight_desc
  inflight_split_region
  inflight_split_desc

Note that it is not a complete support for inflight, additional changes
will be made in the next commits:
  - Turn on the inflight capability for the device
  - Properly initialize virtqueue itself
  - Support resubmit functionality

Signed-off-by: Dima Stepanov <dimastep@yandex-team.ru>