queue.h: rename from intrusive_list.h
It's a trivial (so far, will be extended in a followup patch) wrapper
around standard BSD-originated <sys/queue.h>, no point to name it
differently.

While at this, drop the bogus "#if linux".  First, the header comes from
BSD and is present on many systems, second, it comes with glibc so
should be available on all glibc-based ones, and last, but most
important, it converts the clear error that would result from a missing
header into meaningless "Implement me".