event: add bottom halves
Add the infrastructure to schedule a unit of work in the desired event
loop from outside (another thread).

Following QEMU it's called bottom halves, as it's similar in spirit to
the bottom half concept in operating systems: the units of work are
executed when event loop wakes up and registers the events of interest,
but before the event handlers are dispatched.

The implementation is partially stolen from QEMU, too, modulo the
simplifications done in the assumption of no event loop nesting.