event: avoid unnecessary notifications
If the notifier has already been signaled another notification is
unnecessary and can be skipped.

The same holds on the other side: if there were no notifications there's
no need to read the eventfd.

Use an atomic variable for that.  atomic_xchg is not free but is still
much cheaper than a syscall.