event: simplify event loop cleanup
The code to terminate the event loop and free the associated data is
terribly convoluted for no reason, as it's trying to solve a
non-existing problem of concurrent release and testing for existence
(and it's still racy too).

A much saner approach is to only free the event loop when there's no
concurrent access to it.  One way to do it is to do free at the end of
the thread running the event loop (currently done in the
libvhost-managed global vhost event loop).  Another is to wait for the
thread running the event loop to terminate (to join it) and only do free
afterwards (this is how NBS manages request queues).

Either way this whole mess is unnecessary, so drop it.  For consistency,
move vhd_free_event_loop call to vhd_stop_vhost_server after
pthread_join, to make both users of eventloop use the second approach
above.

Adjust the test accrodingly.