virt_queue: use barrier() for inflight data ordering
The inflight region is never accessed concurrently.  The only ordering
that is necessary for the inflight region beyond the regular vring
ordering is that the used_idx is updated after the corresponding
descriptor, as seen by the next incarnation of the vhost backend
process, regardless of where the current one dies.  For that, only a
compiler barrier is necessary.

Fixes: 1427414 ("virtio/virt_queue: update inflight descriptors for avail and used")