server: only schedule completion BH when necessary
There's no need to schedule the completion BH if the completion being
queued is not the first one in the completion queue, because the first
one must have already done so.

We can get the information on whether there were completions on the
queue before we submitted ours for free, as a result of atomic head
insertion, so this looks like an obvious optimization.