bio: move completion handler to private wrapper
Request completion handler is not anything the library client is
interested in.

Move it from the public API struct vhd_bdev_io to the private wrapper
struct vhd_bio, and adjust the callsites accordingly.

This field is never used outside the library, so it's ok to hide it from
the API.