keep guest memmap while handling the request
The guest memory is accessed directly by the client code.  Therefore
measures should be taken to prevent guest memory from disappearing from
under the client while the request is being processed.

Achieve this by grabbing a reference to the memory map when handing the
request over to the client, and releasing it on receiving it back for
completion.