logging: prepend (again) func:line: to logged messages
As a result of the logging rework, the logged messages are no longer
prefixed with the function name and line number of where the message
originated from.

In addition to missing a useful bit of information at higher logging
levels, this turned VHD_LOG_TRACE into a complete nonsense (which GCC,
when used as the compiler for the project, detects with "zero-length
gnu_printf format string" warning).

Bring func:line: prefix back.  This appears to be useful at least at
this development stage.  When the project stabilizes we may want to
reconsider, but then VHD_LOG_TRACE will need to go altogether.