vdev: don't return error from server event handlers
Errors returned from event handlers are passed on to the event loop and
cause it to terminate.  IOW such errors are treated as fatal to the
whole system.

Protocol errors on the control socket of a particular device are not to
be considered fatal to the whole system.  OTOH they probably are to the
particular connection.

So, instead of error return from the respective event handlers,
uncondintionally return zero (success), but make errors in the handlers
trigger connection termination and transitioning to the listening state.