virtio_blk: fully complete the request on failure
When a request has valid framing (the header with the request
description and the trailer for status), it should be completed fully
regardless of whether it was successful or not.  This means that
fail_request should not abort the request, but rather do the regular
completion with the indices update and notification.

The abort path (when the framing is incorrect and so something went
seriously wrong) is left as is for now, but a log message is added to
indicate the problem.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>