virt_queue: virtio_iov: split in/out segment counts
Per virtio spec,

  2.6.4.2 Driver Requirements: Message Framing
  The driver MUST place any device-writable descriptor elements after
  any device-readable descriptor elements.

This allows certain optimizations and simplifications.

To make use of this property further down the road, maintain separate
counters for device-readable and device-writable segments in virtio_iov,
and verify that no writable descriptors come before readable ones.
Also introduce pointers to the start of the device-readable and
device-writable parts of the descriptor array to ease further accesses
to the respective descriptor ranges.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>