event: change vhd_run_queue interface
Return 0 on running a terminated request queue to indicate
that further queue running is pointless.
Return -EAGAIN if queue is runnable.

This is needed to make the user interface cleaner.
User don't need additional checks to get queue state. It should
run the queue while it returns -EAGAIN, othewise the queue is not
and can't become runnable, so it should be destroyed.

Signed-off-by: Denis Plotnikov <den-plotnikov@yandex-team.ru>