blockdev: add correct serial handling
We used 'id' configuration parameter both as socket path and
disk serial because of unclear member naming and description.
As a result of this confusion we reported chunk of socket path
string to guest as disk serial and there was no way to specify
correct disc serial (without breaking socket path).

This is wrong and disk serial should be configured independently
of socket path.
Replace ambiguous 'id' with two separate explicit options.

Signed-off-by: Anton Kuchin <antonkuchin@yandex-team.ru>