bio: carry vring instead of vdev in bio
We'll need to be able to do per vring operations on a request completion
(e.g. vring draining). We didn't carry information about vring in
a request and carried vdev link only.
Vring has a link to its vdev. So, to satisfy our need replace vdev
link with vring link.

Signed-off-by: Denis Plotnikov <den-plotnikov@yandex-team.ru>