vdev: disable vring in data plane with inflight requests draining
Patch fixes current error-prone vring disabling.

The vring was disabled from the control plane without proper synchronization
with the data plane. This led to differen "races" when a disabled vring was
being processed in data plane.

Disabled vring processing in data plane may result in random bugs:
1. (observed)  inconsistent data sending to master as reply to
   a command, e.g. "last_avail" sending in reply to
   VHOST_USER_GET_VRING_BASE
2. (potential) use-after-free of virtio buffers
3. other unpredictable effects because of the assumption that
   disabled vring isn't processed

The patch fixes it by disabling vring in data plane.

Now, the vring disabling gurantees that:

1. no new requests is processed from avail vring
2. all in-flight requests are completed

Thus, the patch makes the device virtqueues migration-ready because
there is no need to migrate in-flight requests.

Signed-off-by: Denis Plotnikov <den-plotnikov@yandex-team.ru>