event: drop vhd_event_loop_terminated
The flag indicating whether the event loop is being terminated is only
safe to access from the same thread that runs the event loop itself.

Nor is it really needed outside of the event loop.

So drop the function, remove its callsites outside of event.c, and
replace the remaining callsites inside event.c by an open-coded
dereference.

While at this, assert the event loop is terminated when it's being
freed.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>