vdev: use correct name for vring activity state
It turns out that we've mixed up the terms in the code.  We use
"enabled" instead of "started".

According to vhost-user spec, "enabled" controls whether the data is
passed between the ring and the backend, whereas "started" is
responsible for processing the ring itself (i.e. virtio elements
transitioning from avail to used).

"Enabled" property is useful in net-like applications, with the meaning
similar to "being plugged into the network".

The patch renames "enabled" propery to "started" to make the code more
consistent with vhost-user spec.

In vhost-user-blk vrings are always enabled; there's no need in a
separate property controlling it.

Signed-off-by: Denis Plotnikov <den-plotnikov@yandex-team.ru>