vdev: raise error if VHOST_USER_F_PROTOCOL_FEATURES is negotiated
QEMU not has to negotiate VHOST_USER_F_PROTOCOL_FEATURES as it doesn't
implement VHOST_USER_SET_VRING_ENABLE. Moreover, it seems, it will
never implement it as VHOST_USER_SET_VRING_ENABLE is useless in
blk scenarios, since there's no a third party with incomming data which
needs to be enabled or disabled.

This patch adds a check that VHOST_USER_F_PROTOCOL_FEATURES isn't
negotiated and makes some simplification in the code relying on that fact.

Signed-off-by: Denis Plotnikov <den-plotnikov@yandex-team.ru>