vdev: cleanup handling notifier association messages
Handlers for vhost protocol messages that associate various notifier
file descriptors with vrings -- kickfd, callfd, and errfd -- is a mess:
they call into a common function which then does a switch based on the
kind of the notifier.

Unscramble them so that the common part is consolidated in a subroutine
and the specific steps are left in the per-notifier-kind handlers.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>