event: allow to assert being in the home thread of an event loop
A number of operations are unsafe when executed in an arbitrary thread.
Typically they are only safe when performed in a thread running the
corresponding event loop.

In order to allow catching unsafe accesses, introduce the notion of a
home thread for an event loop.  For that, establish the association
between the thread and the event loop on the first iteration of the
latter, and assert it remains this way ever after.

This will also allow to assert this for any other operation that has to
be performed in the thread running the event loop in order to be safe.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>