event: allow to submit work on event loop and wait for it to complete
Introduce a mechanism to submit a work item onto an event loop and wait
for it to complete.  The work may signal completion either directly or
in a delayed callback such that the event loop can make progress while
the work is being done.

As the newly introduced vhd_submit_work_and_wait blocks until the work
is finished, it shouldn't be called in the target event loop, otherwise
it would deadlock.  To ensure such errors are caught early, assert this
in the code.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>