vdev: allow to submit work and wait for it to complete
Introduce a mechanism to submit a work item for a vdev on the control
event loop and wait for it to complete.  The work may signal completion
either directly or via a callback such that the control loop can make
progress while the work is being done.

This is a building block to allow external contexts to trigger state
transitions of the vdev.  As this is the only envisioned use case, only
add a single work slot per vdev, so that no concurrent works are
allowed.

Also introduce the first use of this interface for stopping the vdev.
It's incomplete as the vdev state transitions are not yet properly
consolidated within the control event loop, but it allows to see how the
interface should be used, and simplifies further steps.  Besides, since
this now formally can fail due to the work slot being busy by another
concurrently executing request, change the signature of
vhd_vdev_stop_server to return the status code (the API user doesn't
neeed to immediately update as this is backwards-compatible at the
source level).

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>