rework i/o handler registration scheme
The API to register/unregister i/o event handlers is awkward: it exposes
too much of the internal state, it doesn't clearly define lifetime and
access rules and so on.  This makes it hard to reason about and to use
in a safe race-free manner.

Rework it into a more robust and consistent scheme.  Specifically

- the i/o handler descriptor object becomes opaque to its users accessed
  by pointer only; this prevents unauthorized access to its fields

- it defines (and ensures via assertions) the use of handler descriptors
  that make it less prone to use-after-free errors

- it guarantees that the handler won't be run after it has been
  unregistered

- looks a more natural API overall.

The users of the API are adjusted respectively, too.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>