event: drop atomic from num_events_attached
Now that both attachments and detachments happen in the same event loop,
there's no need for the counter to be atomic.

Make num_events_attached non-atomic, and adjust correspondingly the
places where it's accessed.

While at this, add a note that this counter is only used to check
attach/detach balance.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>