vdev: drop .max_queues
Contrary to what the comment says, the number of queues can not be
modified during the server lifetime.  The only thing that can be changed
is the value of num_queues in the device config space which doesn't
affect anything.

So nuke vhd_vdev.max_queues and use .num_queues universally.  While at
this, also verify that max_queues passed to vhd_vdev_init_server fits
into the smallest size used for the queue index in vhost-user protocol.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>