virt_queue: verify inflight queue size on init
The number of descriptors that can be saved into the per-virtqueue
inflight region is specified in a separate message from the length of
the virtio queue itself (VHOST_USER_{G,S}ET_INFLIGHT_FD for the former
vs VHOST_USER_SET_VRING_NUM for the latter).

Verify that the two are compatible, i.e. that the inflight region is at
least big enough to hold all the virtqueue descriptors, and virtqueue
init, which is done at vring start.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>