decouple dirty log from memory map
The lifetime of the dirty memory log is not tied to that of the memory
map.  Storing the memory log attributes on vhd_memory_map is just wrong,
and results in logging failures as soon as the memory map is changed.

To fix it, encapsulate the memory log attributes in a dedicated
structure, vhd_memory_log, and pass it around into dirty logging
primitives.  Maintain a pointer to it on vdev.  Also link it onto
virtqueues for easier separation of concerns; update these links once
the logging status changes.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>