memlog: abstract out into separate file
Manipulation of the dirty memory log is a fairly big chunk of
functionality.  Keeping it in vdev.c makes it hard to follow and messes
up with the vdev and vring state machine which is already complex enough
in itself.

Abstract out memlog-related stuff into a separate file, with a
reasonably concise interface.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>