memmap: abstract out into separate file
Manipulation of memory map is a fairly big chunk of functionality.
Keeping it in vdev.c makes it hard to follow and messes up with the vdev
and vring state machine which is already complex enough in itself.

Abstract out memmap-related stuff into a separate file, with a
reasonably concise interface.  One extra benefit is that it will allow
to implement support for changing the memory map via
VHOST_USER_ADD_MEM_REG and VHOST_USER_DEL_MEM_REG messages.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>