event: avoid using errno after logging it
Logging functions most certainly call into libc functions so they may
change the errno.  So after logging the value of errno it's no longer
valid, and it's incorrect to use it for the function return value.

Save errno into a local variable and to make sure it's valid throughout
its scope.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>