vdev: suspend receiving messages while processing one
Introduce a pair of functions, vdev_handle_start and vdev_handle_finish,
to bracket the handling of a vhost-user message.

For now all they do is suspend receiving further messages from the
socket while handling the just received one, and resume it afterwards.
This doesn't look terribly useful at the moment because handling of the
message is supposed to be fully contained within the handling function,
and thus the reception is resumed before leaving the epoll callback.
This is going to change in followup patches though so this pair of
functions will make more sense.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>