vdev: don't generate SIGPIPE on sendmsg
If the client closes the connection while the server sends a message,
SIGPIPE is generated which may be detrimental to the server process.

We're better off receiving -EPIPE and handling it.

Ignoring SIGPIPE globally is not an option as this is a library.

Pass MSG_NOSIGNAL flag to sendmsg() to achieve this.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>