vdev: acknowledge messages in their respective handlers
Move sending of acknowledgements to the respective message handler
functions.

This will bring two advantages.  Firstly, it will make the logic of
which message acknowledgement is optional depending on the negotiated
protocol features and the incoming message flags and which is mandatory,
more explicit.

Secondly and more importantly, it will allow to postpone the
acknowledgement until all the side effects of the message are complete,
which may happen well after the handler function returns.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>