vhost_spec: remove packed attribute from message payload structures
Somewhat unexpectedly QEMU doesn't declare vhost-user message payload
packed.  This happens to be an issue for, in particular, inflight
descriptor which is tail-padded when naturally aligned.

Follow suit and drop packed attribute from those structures, to
synchronize the sizes with QEMU.

Also, as the message header and payload are now handled separately, get
rid of struct vhost_user_msg altogether and thus avoid problems with
taking pointers to members of a packed structure.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>