vdev: combine msg_get_vring and msg_valid_num_fds
These two functions are always called in a pair in the same pattern.

Consolidate them into a single function which locates the u64 within the
message payload, extracts the queue number and the invalid fd flag,
finds the respective vring, and validates the number of file descriptors
against the flag.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>