vdev: open-code get_vring_not_started
get_vring_not_started doesn't buy much: it's only used in two callsites.
OTOH it conceals the error location (because VHD_LOG_ERROR includes the
current function, while the really useful one is one level up the
stack).

Open-code it at the callsites instead.  This also allows to return a
more meaningful error code (not that we make any use of it though).

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>