vdev: simplify memory map lifetime management
Memory map lifetime management is fairly convoluted: it's refereced by
vdev, vrings, and in-flight requests.  This makes it hard to track the
references, especially on error paths.

OTOH per-vring references don't look necessary: we just need to make
sure that the old memory map doesn't disappear until all the vrings
switch to the new one.

For this, introduce a new field, vhd_vdev.old_memmap, to keep the
previous memmap untill the switch is complete.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>