vdev: introduce two-stage virtqueue parameter update
Performing non-trivial operations, like updating the virtqueue
parameters in response to vhost-user control messages, and especially
handling errors during those updates, in dataplane threads is
cumbersome.

To make things more conrollable, perfom the update in two stages: first
fully initilize the virtqueue parameters in the control plane, including
consistency checks and error handling, in a shadow copy, such that it
doesn't affect the dataplane; then sync the parameters from the shadow
copy into the actual virtio_virtq, which is a trivial operation that
can't fail.  In the followup patches the second stage will be moved into
the dataplane and thus ensure atomicity.

The shadow copy sits on vhd_vring and is namespaced into a structure.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>