vdev: log requests when starting and finishing handling them
Bringing up a vhost-user connection involves a few dozen steps, and it's
hard to diagnose without knowing which step we're at.

To make diagnostics easier, for every processed vhost-user request log
(at INFO level) its symbolic name and integer value at the beginning and
at the end of handling it.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>