logging: add helpers for object-tagged log messages
In the situation where the vhost server handles multiple devices, with
multiple queues each, it becomes hard to tell which particular object
(among the multiple ones of the same type) a message pertains to.

To make it easier yet generic enough to be usable with all types of
objects, implement another set of logging macros, taking additionally a
pointer to the object the message pertains to.  The object is assumed to
have a const string field ->log_tag which is then prepended to the
message.

The patches introducing the use of these macros will follow.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>