virt_queue: relax inflight state consistency requirements
Normally the inflight region is owned by the client but its content is
fully the server's responsibility.  However, in case of an error on the
client side the consistency checks may fail, and it doesn't look wise to
crash the server (serving multiple clients) just because of this.
Moreover the client may happily continue operation unless the server
restarts under it.

So replace the assertions with warnings.

Note that the original problem isn't guest-triggerable so it doesn't
look like having a DoS potential.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>