cancel queued requests on vring disconnect
When the library user forces a disconnect of a client, (s)he's likely to
expect that, once vhd_vdev_stop_server() returns, no more requests will
reach the backend.

vring_stop_bh() disconnects the kick notification which indeed ensures
no more virtio queue processing will happen.

However, there may already be requests extracted from the virtio queue
and queued on the request queue for the backend to consume.

To ensure those requests won't reach the backend, introduce a function,
vhd_cancel_queued_requests(), which filters the requests on the request
queue originating from the given vring and cancels them, and call this
function when the vring is being stopped for disconnection.

To that end, update the comment on vhd_vdev_stop_server().

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>