add CANCELED status for queued requests
When the server prepares to disconnect the client, it doesn't have to
wait until all the queued requests are completed: the client is supposed
to make use of the inflight region, and therefore be ready for the
connection to be dropped abruptly; the library will then resubmit the
requests that hadn't completed in the previous session.

However, such requests need to relinquish the allocated resources, drop
the references taken, and so on.  IOW they need to go through the
regular completion procedure, except that they shouldn't be delivered to
the virtio used ring.

Introduce a dedicated status for such requests, VHD_BDEV_CANCELED, that
makes the completion procedure skip updating the vring.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>