virt_queue: consolidate checks for descriptor chain length
The (only) right place to check for descriptor chain length to not
exceed the queue size is map_buffer().

Furthermore, when failed, it doesn't necessarily mean a descriptor loop:
e.g. the driver may have created an indirect descriptor of excessive
length or what not.

Adjust the error message and propagate the error from this function up
the stack at all callsites.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>