virt_queue: fix unsigned arithmetic in inflight handling
If the server dies after updating the used ring but before adjusting the
inflight region, a few entries may remain marked inflight while they
actually aren't.  To compensate for that, those entries are forced to
inflight=0 on reconnection.

The count of those entries is obtained as the difference between the
actual index in the used ring and the used_idx value recorded in the
inflight region.  Both of these are uint16_t.  However, their difference
is assigned to an int variable, which is then used as a decremented
counter to traverse the chained entries of the last batch until it
reaches zero.

As a result, when the used index value overflows but used_idx still
doesn't, the resulting integer counter becomes negative, and the last
batch chain traversal goes far beyond its length.  For many of those
entries the linking field .next may be stale as they do not participate
in the last batch.

Make the variable uint16_t.  Also take into account that we actually
don't batch used ring updates and therefore should never have batch size
bigger than one, so assert that.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>