virt_queue: don't clean inflight entries upon migration
When a connection is established by a qemu accepting incoming migration,
the used_idx in a freshly initialized inflight region doesn't
necessarily match the the used ring index in the guest memory as it
may have already transferred some descriptors.

This condition is however mistreated by virtq_inflight_reconnect_update
as an indication that some inflight entries have incorrectly been left
in "inflight" state after having been exposed as used to the guest, as
may happen on accidental disconnects.  It then starts to reset those
excessive inflight entries.

This makes no real harm with the current code as in reality it doesn't
change the zeroed state of the inflight region, but is still logically
wrong.

Detect the situation of a clean inflight region and skip resetting
inflight entries in this case.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>