virt_queue: split freeing the element from pushing it to used ring
virtq_commit_buffers does two things: it pushes the virtq element onto
the used ring for the driver to consume, and frees the resources
associated with the element.

However, for a special case when the element processing is canceled, the
element shouldn't be exposed as used to the guest, the idea being that
this scenario happens on a voluntary disconnection during graceful
server shutdown, and the processing will pick up upon restart at where
it left off with the help of vhost-user inflight machinery, if desired.

This currently results in a leak of a virtq element and a reference to
the memory map, which makes the client (i.e. guest) memory stick in the
server's memory even after the client disappears.

To that end, split pushing the element onto used ring and freeing it.
Adjust the callers accordingly, in particular, free the element event if
pushing it to the used ring is to be skipped.

Also, to make the function name better represent what it does, and to
make it easier to locate the callsites that need adjustment, rename
virtq_commit_buffers to virtq_push.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>