virt_queue: merge memmap unref into virtio_free_iov
Drop the reference to the memmap associated with the virtq element in
the same function that frees the element itself.

In addtion to these two actions being more logical as a single step,
this change paves the way to the followup patch where this step is
performed separately on certain paths.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>