virtio/virt_queue: ensure inflight update ordering
The inflight region is not accessed concurrently, so there's no need for
smp memory barriers when updating it.  However, it has to be updated in
a fixed order, so that no matter when the server dies, the inflight
region can be predictably used to restore the virtio device operation.

To ensure that, add a couple of compiler barriers.  They fix no known
bugs with the compilers used, so this is just a precautionary measure.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>