vdev: close the control connection last at cleanup
When the server for a device is being gracefully stopped, the control
connection is closed before the client mappings (guest memory, inflight
region, dirty memory log) are detached.  This creates a window when the
client may notice the connection being broken and start a new
connection, potentially with a different server (who might have
taken over the server socket).  As a result, those mappings may turn out
mapped in both the old and the new server processess, with all sorts of
possible damage.

To prevent that, don't close the connection untill all other links to
the client are broken.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>