vdev: keep old callfd/errfd while pushing new one to dataplane
SET_VRING_CALL and SET_VRING_ERR control messages may (and often do)
arrive when the vring is already started.  As the respective message
handlers close the old descriptor immediately to replace it with the
new one, the copy of the descriptor for the dataplane use becomes stale,
until the new one is synced there.  During this window the dataplane may
happen to still use it, e.g. to write into it for notification.  Chances
are that this file descriptor gets reused for something else (e.g. the
new client control connection, the guest memory map, etc.); in that case
those writes may cause real damage (mess the wire protocol on the
socket, induce guest memory corruption, etc.).

To prevent this from happening, keep the file descriptor open until the
new one is propagated to the dataplane and message handling is
completed.  For that, introduce a field on vdh_vdev to hold a copy of
this descriptor during the controlplane -> dataplane -> controlplane
roundtrip.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>