vdev: avoid multiple disconnect on errors
Being the best (or, rather, the only feasible) strategy to deal with
errors on the control connection, vdev_disconnect is called on a few
error paths.  Some of the error paths can be taken with the connection
already closed due to an earlier vdev_disconnect.

However, vdev_disconnect doesn't survive being called on an already
closed connection and crashes.

Prevent this by checking if the connection is still alive, and bailing
out early if it isn't.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>