virtio/virt_queue: allow descriptor chains bigger than queue size
According to virtio spec, "a driver MUST NOT create a descriptor chain
longer than the Queue Size of the device", even when indirect
descriptors are used.

Windows virtio drivers, both viostor (virtio-blk) and vioscsi
(virtio-scsi) violate this requirement: they hardcode the value that
defines the maximum length of the descriptor chain to be 512 + 3.

As a result, when sufficiently large i/o happens, the driver may submit
a request with a descriptor chain bigger than the queue size, which
libvhost (rightfully) rejects.

Support such broken drivers by maintaining a separate field for the
maximum allowed chain length, which is set to 512 + 3 or the queue size,
whichever is bigger.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>