virtio/virt_queue: use memory barrier in !EVENT_IDX case too
If VIRTIO_F_RING_EVENT_IDX feature is not negotiated, the driver may
still attempt to optmimize the rate of interrupts on the used ring by
setting VIRTQ_AVAIL_F_NO_INTERRUPT flag in the avail ring.

Reading this flag must not happen before the used ring is exposed to the
driver, so a full memory barrier is needed, too, just as in EVENT_IDX
case.

So do a memory barrier regardless of whether EVENT_IDX feature is
enabled.

For better readability, factor out the check for the notification being
needed by the driver into a separate function.

Fixes: bf912a0 ("virtio/virt_queue: check queue flags before notifying the guest")
Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>