vdev: log number of inflight requests at vring stop
Knowing how many requests were in progress when the virtqueue was
requested to stop may help debugging stalls etc.

Log this number when processing the queue stop request.  To be safe
against long / blocking operations in the logging functions, save this
number in the dataplane event loop into a copy and log it in the control
event loop.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>