vdev: remember message being handled on vdev
There can only be at most one message being handled by the vdev state
machine, so there's no need to pass around the message opcode.

Just store it on vdev on entry into the handling, and reset to
VHOST_USER_NONE on exit, and use the stored value throughout.

This also paves the way to refer to the message currently being handled
when logging.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>