virtio_spec: update virtq_used_elem.len comment
Virtio spec v1.1- describes the .len field in struct virtq_used_elem in
a confusing way.  This (presumably) resulted in libvhost filling this
field incorrectly: with the number of used descriptors in the chain
rather than the total byte length of the bytes written in the used
chain.

Update the wording of the comment for this field as documented in the
spec v1.2, which is more explicit and clear.

Adjustments to the handling of this field will come in the followup
patches.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>