virtio: pass proper len in virtq_used_elem
The field .len in virtq_used_elem is incorrectly interpreted as the
number of descriptors in the chain.

In reality this field is supposed to contain

  The number of bytes written into the device writable portion of the
  buffer described by the descriptor chain.

as now worded explicitly in virtio spec v1.2 section 2.7.8.

However traditionally this field was used very inconsistently, and so no
driver even cared to look at it.

Until virtiofs driver for Windows did, and this is where things break.

To address this problem, let the device code pass the number of bytes
written in an argument to virtq_push.  Unfortunately, figuring out the
value to pass, without major rework to split OUT and IN buffers, is
clumsy.  Do it correctly but hackishly for virtiofs, and simply pass
zero for virtio-blk where no known drivers seem to care.

Update the tests accordingly.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>