virtio_fs: don't verify virtio_fs_in_header.len against buffer size
Certain guest driver versions, in particualr, windows driver until
commit 5255fd2e "VirtIOFS: RHBZ#2024518: fix FUSE_READ input buffer
size", pass incorrect .len value in virtio_fs_in_header, which causes
libvhost's virtiofs handler to abort such requests and therefore
malfunction.

To work around this problem, skip the check for equality of this value
to the size of the OUT (i.e. fuse-IN) buffer, and rely on the backend to
do the proper checks.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>