virtio_fs_spec: add comment about duplicate definitions
virtio_fs_spec contains definitions of virtio_fs_in_header and
virtio_fs_out_header, which actually aren't defined in the virtio spec,
but are rather duplicates of the respective fuse_in_header and
fuse_out_header structures defined in the FUSE spec.

Add a comment documenting this, and the desire to remove the duplicates
in favor of the originals.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>