virtio_fs: drop redundant field from virtio_fs_io
Now that IN and OUT buffers are separate there's no need to record the
start of the IN buffers in virtio_fs_io in order to be able to locate
the fuse_out_header on completion and extract the used virtq element
written length.

Use the start of the IN buffers directly from virtio_iov.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>