virtio_blk: properly fill used virtq element written length
Now that IN and OUT buffers are separate it's easy to figure out the
used virtq element written length: for virtio-blk, where the last byte
of the IN buffer is to contain the request completion status, the whole
IN buffer is counted as written, so its total length should be returned
in the length field of the used element.

Replace the ugly hack where we pass zero here in the hope that no guest
driver actually checks it with the correct value.

Signed-off-by: Roman Kagan <rvkagan@yandex-team.ru>